peliculas_proyecto/requirements.txt
streamlit
pandas
Pillow
seaborn
matplotlib
wordcloud
